translation-verifier-ui-app/requirements.txt
Flask
openai
python-docx
